UI, AI, script fix
Applied Dave's UI
Fixed the issue of Azhang not spawning after the event
Now Endless Tomb AI is 'always_use_only_land'. Now AI summons the endless tomb at the start of the battle when possible.
Changed Nagash's set effect to the cataclysm death spell, and also changed the effect enabling as it was in the wrong table.